+++ index.js
@@ -4,3 +4,4 @@ exports.arrayset = require('./lib/arrayset');
 exports.object = require('./lib/object');
 exports.objectset = require('./lib/objectset');
 exports.datetime = require('./lib/datetime');
+exports.predicate = require('./lib/predicate');