+++ RELEASENOTES.md
@@ -4,6 +4,12 @@
 -objectset.deepValue(obj, deepProperty, value)
 -predicate.hasProperty
 -predicate.hasNoProperty
+-predicate.inValueSet
+-predicate.hasNoProperties
+-predicate.propertyValue
+-predicate.falsyPropertyValue
+-predicate.or
+-predicate.and
 
 # v1.4.0 - February 27, 2015