+++ RELEASENOTES.md
@@ -1,3 +1,16 @@
+# v1.4.1 - TBD
+
+-object.deepValue(obj, deepProperty);
+-objectset.deepValue(obj, deepProperty, value)
+-predicate.hasProperty
+-predicate.hasNoProperty
+-predicate.inValueSet
+-predicate.hasNoProperties
+-predicate.propertyValue
+-predicate.falsyPropertyValue
+-predicate.or
+-predicate.and
+
 # v1.4.0 - February 27, 2015
 
 This is the initial release of blue-button-util.js library.  It includes javascript utility methods.